renamed 'run_test.js' to 'runTest.js' and updated imports
